rename: Number 라는 포괄적인 네이밍을 직관적으로 CardNumber로 수정
