+++ src/Card/ui/Card/Card.stories.tsx
@@ -2,9 +2,9 @@ import { Meta, StoryObj } from "@storybook/react";
 import { CardSize, CardSizeVariants } from "../../utils/style";
 import Card from "./Card";
 import { CardNameProps } from "./core/CardCompany";
+import { NumberProps } from "./core/CardNumber";
 import { ExpirationDateProps } from "./core/ExpirationDate";
 import { NameProps } from "./core/Name";
-import { NumberProps } from "./core/Number";
 
 const meta: Meta<typeof Card> = {
 	component: Card,
@@ -16,7 +16,7 @@ export default meta;
 interface StoryArgs {
 	size: CardSize;
 	"CardCompany.text": CardNameProps["text"];
-	"Number.text": NumberProps["text"];
+	"CardNumber.text": NumberProps["text"];
 	"Name.text": NameProps["text"];
 	"ExpirationDate.month": ExpirationDateProps["month"];
 	"ExpirationDate.year": ExpirationDateProps["year"];
@@ -29,7 +29,7 @@ const CardRender = (args: StoryArgs) => {
 		<Card size={args.size}>
 			<Card.CardCompany text={args["CardCompany.text"]} />
 			<Card.Chip />
-			<Card.Number text={args["Number.text"]} />
+			<Card.CardNumber text={args["CardNumber.text"]} />
 			<Card.Name text={args["Name.text"]} />
 			<Card.ExpirationDate
 				month={args["ExpirationDate.month"]}
@@ -44,7 +44,7 @@ export const CardStory: Story = {
 	args: {
 		size: "small",
 		"CardCompany.text": "동현 카드",
-		"Number.text": "1234 5678 9101 1121",
+		"CardNumber.text": "1234 5678 9101 1121",
 		"Name.text": "장동현",
 		"ExpirationDate.month": "12",
 		"ExpirationDate.year": "23"
