+++ src/Card/ui/Card/Card.tsx
@@ -10,10 +10,10 @@ import {
 	CardTop
 } from "../styles/Card";
 import CardCompany from "./core/CardCompany";
+import CardNumber from "./core/CardNumber";
 import Chip from "./core/Chip";
 import ExpirationDate from "./core/ExpirationDate";
 import Name from "./core/Name";
-import Number from "./core/Number";
 import { CardProvider } from "./provider/CardProvider";
 
 interface CardProps {
@@ -24,7 +24,7 @@ interface CardProps {
 const Card = ({ size, children }: CardProps) => {
 	const cardNameComponent = findComponentFromChildren(children, CardCompany);
 	const chipComponent = findComponentFromChildren(children, Chip);
-	const cardNumberComponent = findComponentFromChildren(children, Number);
+	const cardNumberComponent = findComponentFromChildren(children, CardNumber);
 	const nameComponent = findComponentFromChildren(children, Name);
 	const expirationDateComponent = findComponentFromChildren(
 		children,
@@ -53,7 +53,7 @@ Card.CardCompany = CardCompany;
 Card.Chip = Chip;
 Card.Name = Name;
 Card.ExpirationDate = ExpirationDate;
-Card.Number = Number;
+Card.CardNumber = CardNumber;
 
 export default Card;
