+++ src/Card/ui/Card/EmptyCard.stories.tsx
@@ -1,9 +1,9 @@
 import { Meta, StoryObj } from "@storybook/react";
 import EmptyCard, { EmptyCardProps } from "./EmptyCard";
 import { CardNameProps } from "./core/CardCompany";
+import { NumberProps } from "./core/CardNumber";
 import { ExpirationDateProps } from "./core/ExpirationDate";
 import { NameProps } from "./core/Name";
-import { NumberProps } from "./core/Number";
 
 const meta: Meta<typeof EmptyCard> = {
 	component: EmptyCard,
@@ -15,7 +15,7 @@ export default meta;
 interface StoryArgs {
 	mode: EmptyCardProps["mode"];
 	"CardCompany.text": CardNameProps["text"];
-	"Number.text": NumberProps["text"];
+	"CardNumber.text": NumberProps["text"];
 	"Name.text": NameProps["text"];
 	"ExpirationDate.month": ExpirationDateProps["month"];
 	"ExpirationDate.year": ExpirationDateProps["year"];
@@ -28,7 +28,7 @@ const EmptyCardRender = (args: StoryArgs) => {
 		<EmptyCard mode={args.mode} onClick={() => {}}>
 			<EmptyCard.CardCompany text={args["CardCompany.text"]} />
 			<EmptyCard.Chip />
-			<EmptyCard.Number text={args["Number.text"]} />
+			<EmptyCard.CardNumber text={args["CardNumber.text"]} />
 			<EmptyCard.Name text={args["Name.text"]} />
 			<EmptyCard.ExpirationDate
 				month={args["ExpirationDate.month"]}
@@ -43,7 +43,7 @@ export const EmptyCardStory: Story = {
 	args: {
 		mode: "add",
 		"CardCompany.text": "",
-		"Number.text": "",
+		"CardNumber.text": "",
 		"Name.text": "NAME",
 		"ExpirationDate.month": "MM",
 		"ExpirationDate.year": "YY"
